# backend/public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "O'Secours"
  description: 'API pour gérer des ressources'
  version: '1.0'
servers:
  -
    url: 'http://localhost:8000/api'
    description: 'API Server'
paths:
...
  '/documents/{id}':
    get:
      tags:
        - Documents
      summary: 'Get a specific document'
      description: 'Returns a single document by ID'
      operationId: 9cf5289c9ad15ddb6da247477998f54b
      parameters:
        -
          name: id
          in: path
          description: 'Document ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: 'Document not found'
    put:
      tags:
        - Documents
      summary: 'Update a document'
      description: 'Updates an existing document'
      operationId: 8340a1a99f82c944b6c7b45b0794505f
      parameters:
        -
          name: id
          in: path
          description: 'Document ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                filename:
                  type: string
                description:
                  type: string
                file:
                  type: string
                  format: binary
              type: object
      responses:
        '200':
          description: 'Document updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: 'Document not found'
    delete:
      tags:
        - Documents
      summary: 'Delete a document'
      description: 'Deletes a specific document'
      operationId: dbfb02aca266743c29805da7424360eb
      parameters:
        -
          name: id
          in: path
          description: 'Document ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Document deleted'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Document deleted successfully.' }
                type: object
        '404':
          description: 'Document not found'
  '/documents/store/animals/{animalId}':
    post:
      tags:
        - Documents
      summary: 'Add a document to an animal'
      description: 'Creates a new document associated with an animal'
      operationId: 40c0c96e0c8c87f98a6dfba899f3f6e8
      parameters:
        -
          name: animalId
          in: path
          description: 'Animal ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - filename
                - file
              properties:
                filename:
                  type: string
                description:
                  type: string
                file:
                  type: string
                  format: binary
              type: object
      responses:
        '201':
          description: 'Document created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: 'Invalid input'
  '/animals/{animalId}/documents':
    get:
      tags:
        - Documents
      summary: 'Retrieve all documents associated with an animal'
      description: 'Fetches a list of documents associated with a specific animal'
      operationId: 575fb5f5664c376a3fc31ffdacb60772
      parameters:
        -
          name: animalId
          in: path
          description: 'Animal ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
...
  '/shelters/{shelterId}/documents':
    get:
      tags:
        - Documents
      summary: 'Retrieve all documents associated with a shelter'
      description: 'Fetches a list of documents associated with a specific shelter'
      operationId: cf6ec6e6bb588a2c30a33aa5b8cf2d94
      parameters:
        -
          name: shelterId
          in: path
          description: 'Shelter ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
    post:
      tags:
        - Documents
      summary: 'Add a document to a shelter'
      description: 'Creates a new document associated with a shelter'
      operationId: c236d47f75fdcf7f8ceda4df7cd2eac7
      parameters:
        -
          name: shelterId
          in: path
          description: 'Shelter ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - filename
                - file
              properties:
                filename:
                  type: string
                description:
                  type: string
                  nullable: true
                file:
                  type: string
                  format: binary
              type: object
      responses:
        '201':
          description: 'Document created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: 'Invalid input'
components:
  schemas:
...
    Document:
      title: Document
      description: 'Document model'
      properties:
        id:
          description: 'ID of the document'
          type: integer
          example: 1
        filename:
          description: 'Name of the file'
          type: string
          example: document.pdf
        description:
          description: 'Description of the document'
          type: string
          example: 'A document describing the health status'
        size:
          description: 'Size of the document in bytes'
          type: integer
          example: 1024
        url:
          description: 'URL of the document'
          type: string
          example: 'http://example.com/document.pdf'
        date:
          description: 'Date of the document'
          type: string
          format: date
          example: '2024-04-16'
        mimetype_id:
          description: 'ID of the MIME type'
          type: integer
          example: 1
        doctype_id:
          description: 'ID of the document type'
          type: integer
          example: 1
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'JWT Authorization header using the Bearer scheme.'
      bearerFormat: JWT
      scheme: bearer